move view related files into lib/views
