+++ frontend/viewer/src/lib/OverrideFields.svelte
@@ -1,6 +1,6 @@
 ﻿<script lang="ts">
 
-  import {initView, useCurrentView} from '$lib/services/view-service';
+  import {initView, useCurrentView} from '$lib/views/view-service';
   import type {FieldIds} from '$lib/entry-editor/field-data';
 
   export let shownFields: FieldIds[] = [];
