+++ frontend/viewer/src/lib/entry-editor/field-editors/MultiFieldEditor.svelte
@@ -4,7 +4,7 @@
   import { createEventDispatcher } from 'svelte';
   import type { IMultiString } from '$lib/dotnet-types';
   import type {WritingSystemSelection} from '../../config-types';
-  import {useCurrentView} from '../../services/view-service';
+  import {useCurrentView} from '$lib/views/view-service';
   import {useWritingSystemService} from '../../writing-system-service';
   import {makeHasHadValueTracker} from '$lib/utils';
