+++ frontend/viewer/src/lib/entry-editor/field-editors/MultiOptionEditor.svelte
@@ -4,7 +4,7 @@
   /* eslint-disable @typescript-eslint/no-duplicate-type-constituents, @typescript-eslint/no-redundant-type-constituents */
   import {createEventDispatcher} from 'svelte';
   import type {WritingSystemSelection} from '../../config-types';
-  import {useCurrentView} from '../../services/view-service';
+  import {useCurrentView} from '$lib/views/view-service';
   import MapBind from '../../utils/MapBind.svelte';
   import {useWritingSystemService} from '../../writing-system-service';
   import FieldTitle from '../FieldTitle.svelte';
