+++ frontend/viewer/src/lib/entry-editor/field-editors/SingleFieldEditor.svelte
@@ -2,7 +2,7 @@
   import FieldTitle from '../FieldTitle.svelte';
   import type {WritingSystemSelection} from '../../config-types';
   import CrdtTextField from '../inputs/CrdtTextField.svelte';
-  import {useCurrentView} from '../../services/view-service';
+  import {useCurrentView} from '$lib/views/view-service';
   import {useWritingSystemService} from '../../writing-system-service';
   import {makeHasHadValueTracker} from '$lib/utils';
