+++ frontend/viewer/src/lib/entry-editor/object-editors/ExampleEditor.svelte
@@ -3,7 +3,7 @@
   import EntityEditor from './EntityEditor.svelte';
   import MultiFieldEditor from '../field-editors/MultiFieldEditor.svelte';
   import SingleFieldEditor from '../field-editors/SingleFieldEditor.svelte';
-  import {objectTemplateAreas, useCurrentView} from '../../services/view-service';
+  import {objectTemplateAreas, useCurrentView} from '$lib/views/view-service';
 
   export let example: IExampleSentence;
   export let readonly: boolean;
