+++ frontend/viewer/src/lib/views/project-view-state-service.ts
